requirements.txt
flask==3.0.2
flask-socketio==5.3.6
python-dotenv==1.0.0
web3==6.15.1
eth-account==0.10.0
eventlet==0.35.2 